third_party/blink/web_tests/platform/mac-mac10.15/external/wpt/webrtc-stats/stats-caching-expected.txt

This is a testharness.js-based test.
PASS onicecandidate does not cache results
FAIL addIceCandidate does not cache results assert_equals: expected 0 but got 1
Harness: the test ran to completion.
